support logstash 5.x (#5)
* s/plugin/logstash-plugin/g

`plugin` is deprecated. recent logstash documentation refer to
`logstash-plugin`.

while here remove `when: ansible_os_family == 'FreeBSD'`.

* logstash-plugin is also installed without exec bit

* support ubuntu

* introduce logstash_package_name

because FreeBSD has multiple version of logstash

* try logstash 5.x

I cannot test this locally because:

* logstash 5.x is not in the official package tree
* the local package can be retrieved over VPN
* but my local VPN client keeps crashing after a while
* SSH proxy works but a Java package `logstash` depends on does not
  respect HTTP_PROXY

"HTTP_PROXY and HTTPS_PROXY environment variables are not correctly
respected when jar-dependency fetch the pom"
https://github.com/elastic/logstash/issues/6044#issuecomment-253854228

"Proxy support not working if using the environment"
https://github.com/mkristian/jar-dependencies/issues/16
https://github.com/mkristian/jar-dependencies/issues/16#issuecomment-266659534

* enable DEBUG

* more debug flags

* mount fdescfs and procfs

also, remove DEBUG as it turns out that sysutils/logstash5 has a bug
that removes exec bit of bundled gems' binaries. the port has been
fixed.

* the port has been merged

* use canonical port name

* s/fdesc/procfs/

* support logstash5

* wait for logstash after starting it

in the first ansible play, it restarts logstash immediately after
starting. this probably causes a race.

* the default port is 9600

* validate logstash configuration files

as there are multiple files, you cannot use `validate:`.

* a typo

* reload logstash instead of restarting it

when the change is something in configuration files.

* swap single quote and double quote

* remove TODOs

* pipeline.unsafe_shutdown can be enabled in logstash.yml
* visualvm should be handled by its own role
* config.reload.automatic can be enabled in logstash.yml
* reload handler has been implemented
* with ansible-role-redis, you can implement queuing
* plugins has been managed by the role

* comment why pif file is not used

* always run logstash as user logstash

as logstash creates directory and files when invoked.

if logstash binaries, including logstash-plugin, runs as root, it causes
subsequent logstash run to fail.

* fix wrong keys

* use logstash_config_merged instead of logstash_config

* wait for logstash to start in handler

there is a race when HUP terminates restart

* fix a typo

* document the finding

* manage jvm.options

* add other defaults

* document new valiables

* add License

* add some comments

* add comments on the task and the template

with new configuration option system, the task and the template will be
removed

* update the example

* add some notices

* restore logstash_config

no, rc.conf will not be removed as 'conf.d' must bedefined.